Fix integer overflow in PWM current calc
There is an integer overflow in the
PwmGenerationBase::GetCurrent() method. This takes a
uint16_t sample of the current then scales it and returns a
fixed point value for later calculations. The FP_FROMINT()
macro shifts by 5 bits typically which will easily cause and
overflow of a uint16_t value. We cast to int32_t before
converting to fixed point to avoid this.

Tests:
 - Run pwmsimulator on Linux and C2000 and verify the
   output is now identical.

+++ include/pwmgenerationbase.h
@@ -206,7 +206,7 @@ protected:
 
     static s32fp GetCurrent(uint16_t input, s32fp offset, s32fp gain)
     {
-        s32fp il = FP_FROMINT(input);
+        s32fp il = FP_FROMINT((int32_t)input);
         il -= offset;
         return FP_DIV(il, gain);
     }
